Lorentz.Contracts.Upgradeable.Client

Description

Operations related to upgradeable contracts.

Documentation

data UStoreValueUnpackFailed

Failed to code UStore value to given type.

Constructors

UStoreValueUnpackFailed ByteString Text

data UStoreElemRef

Constructors

UrField MText

UrSubmap MText (SomeConstrainedValue PackedValScope')

readContractUStore :: forall v m. (UnpackedValScope v, HasTezosNodeRpc m) => Address -> UStoreElemRef -> m (Value v)

Read UStore value of given contract.

This essentially requires contract having only one big_map bytes bytes in storage.

readContractUStoreEntrypoint :: HasTezosNodeRpc m => Address -> MText -> m [ExpandedOp]

Read an UStore entrypoint. For contracts which are filled with storage-driven approach.

Unlike readContractUStore, here we don't need to know exact type of value (lambda) in order to unpack it, thus returning code in untyped representation.