Lorentz.Contracts.Upgradeable.Common.Base

Documentation

newtype Version

Version of a contract.

Our current versioning suggests that this type is a term-level reflection of types which have KnownContractVersion instance, so this version item should uniquely identify storage structure and entrypoints set for a given contract for all of its instances.

The old semantics of this type was that it counts number of given contract instance upgrades, so different contract instances, being upgraded to the recent version, could have different Versions. For old contracts we have to follow this behaviour.

Constructors

Version

Fields

unVersion :: Natural

type VersionKind = ContractVersionTag -> Type

Kind of type-level contract version.

class KnownContractVersion (v :: VersionKind) where

Declare given type as contract version identifier.

Instances of this typeclass (versions) uniquely identify contract storage scheme and code. Normally the opposite should also hold, i.e. contract version - (contract storage scheme, code) relation is a bijection.

If as part of migration you need to update contract storage without modifying its structure, then contract version should not change, and you should perform an upgrade to the same version as the current one.

We allow upgrades between arbitrary two versions, so one can not only upgrade to the next adjacent version, but also upgrade a new contract from V0 to the recent version immediately, or leave version the same (as a versatile way to change storage).

Minimal complete definition

Nothing

Associated Types

type VerInterface v :: [EntryPointKind]

List of entrypoints of given contract version.

type VerUStoreTemplate v :: Type

Storage template of given contract version.

type VerPermanent v :: Type

Set of permanent entrypoints (as a sum type).

We tie this type to contract version for convenience, in order not to carry one more type argument everywhere. We do not ensure right here that all versions of a contract have the same permanent entrypoints, but if this does not hold, then (ideally) it will not be possible to construct migration between such contract versions.

Methods

contractVersion :: Proxy v -> Version

Get term-level contract version. Returned value will be stored within the contract designating the current contract version.

contractVersion :: (v ~ cid ver, KnownNat ver) => Proxy v -> Version

Get term-level contract version. Returned value will be stored within the contract designating the current contract version.

type VerParam v = UParam (VerInterface v)

type VerUStore v = UStore (VerUStoreTemplate v)

data EmptyContractVersion (perm :: Type) :: VersionKind

Contract with empty interface and storage.

data SomeContractVersion (perm :: Type) :: VersionKind

Version which forgets about particular interface/storage.

type UStore_ = UStore SomeUTemplate

UStore with hidden template.

newtype MigrationScript oldStore newStore

Code of migration for UStore.

Invariant: preferably should fit into op size / gas limits (quite obvious). Often this stands for exactly one stage of migration (one Tezos transaction).

Constructors

MigrationScript

Fields

unMigrationScript :: Lambda UStore_ UStore_

type MigrationScriptFrom oldStore = MigrationScript oldStore SomeUTemplate

Corner case of MigrationScript with some type argument unknown.

You can turn this into MigrationScript using checkedCoerce.

type MigrationScriptTo newStore = MigrationScript SomeUTemplate newStore

data UContractRouter (ver :: VersionKind)

Keeps parameter dispatching logic.

Constructors

UContractRouter

Fields

unUContractRouter :: Lambda (VerParam ver, VerUStore ver) ([Operation], VerUStore ver)

type SomeUContractRouter = UContractRouter (SomeContractVersion ())

data UpgradeableEntryPointsKind

Common marker for upgradeable, or virtual, entrypoints. Can be used when each upgradeable entrypoint is simple, i.e. does not itself consist of multiple entrypoints.

newtype PermanentImpl ver

Implementation of permanent entrypoints.

This will be injected into contract storage as one of fields, so make sure that code within does not exceed several instructions; an actual entrypoint logic can be put into UStore and called from within PermanentImpl only when necessary.

Regarding documentation - this have to provide code pieces wrapped into DEntryPoint with PermanentEntryPointsKind, so always use entryCase as implementation of this type or inject documentation of code which does so unless you know what you are doing.

Constructors

PermanentImpl

Fields

unPermanentImpl :: Entrypoint (VerPermanent ver) (VerUStore ver)

type SomePermanentImpl perm = PermanentImpl (SomeContractVersion perm)

data PermanentEntryPointsKind

Common marker for permanent entrypoints. Can be used when parameter for permanent entrypoints is flat, i.e. does not have nested subparameters with multiple entrypoints.

mkUContractRouter :: ([VerParam ver, VerUStore ver] :-> '[([Operation], VerUStore ver)]) -> UContractRouter ver

coerceUContractRouter :: (Coercible_ (VerParam s1) (VerParam s2), Coercible_ (VerUStore s1) (VerUStore s2)) => UContractRouter s1 -> UContractRouter s2

emptyPermanentImpl :: VerPermanent ver ~ Empty => PermanentImpl ver

Common implementation of permanent part in case contract has no such.

mkSmallPermanentImpl :: forall ver dt out inp clauses. (CaseTC dt out inp clauses, DocumentEntryPoints PermanentEntryPointsKind dt, dt ~ VerPermanent ver, inp ~ '[VerUStore ver], out ~ ContractOut (VerUStore ver)) => IsoRecTuple clauses -> PermanentImpl ver

Construct implementation of permanent part in a common case; this works similarly to entryCase.

Use this function only for very small implementations.

Re-exports

data Nat

(Kind) This is the kind of type-level natural numbers.

data Empty

Replacement for uninhabited type.

absurd_ :: (Empty ': s) :-> s'

Witness of that this code is unreachable.