Lorentz.Contracts.Upgradeable.Common.Contract

Description

Template for upgradeable contract.

It provides the following features: 1. Contract with upgradeable storage format and entrypoints set. 2. Two way to upgrade the contract - one shot and entrypoint-wise.

Documentation

data Parameter (ver :: VersionKind)

Parameter of upgradeable contract. It contains, among others:

Entrypoint for running one of upgradeable entrypoints.

Entrypoint for running one of permanent entrypoints, suitable e.g. for implementing interfaces. 3a. Entrypoint for upgrade in a single call. 3b. Entrypoints for entrypoint-wise upgrade.

Constructors

Run (VerParam ver)

RunPerm (VerPermanent ver)

Upgrade (OneShotUpgradeParameters ver)

GetVersion (View () Version)

SetAdministrator Address

EpwBeginUpgrade ("current" :! Version, "new" :! Version)

EpwApplyMigration (MigrationScriptFrom (VerUStoreTemplate ver))

EpwSetCode SomeUContractRouter

EpwSetPermCode (SomePermanentImpl (VerPermanent ver))

EpwFinishUpgrade

type UTAddress ver = TAddress (Parameter ver)

type UContractRef ver = ContractRef (Parameter ver)

type PermConstraint ver = (NiceParameterFull (VerPermanent ver), NoExplicitDefaultEntryPoint (VerPermanent ver), HasTypeAnn (VerPermanent ver), RequireAllUniqueEntryPoints (Parameter ver))

Constraint on abstract set of permanent entrypoints.

data Storage (ver :: VersionKind)

type UpgradeableContract ver = ContractCode (Parameter ver) (Storage ver)

data PermanentImpl ver

Implementation of permanent entrypoints.

This will be injected into contract storage as one of fields, so make sure that code within does not exceed several instructions; an actual entrypoint logic can be put into UStore and called from within PermanentImpl only when necessary.

Regarding documentation - this have to provide code pieces wrapped into DEntryPoint with PermanentEntryPointsKind, so always use entryCase as implementation of this type or inject documentation of code which does so unless you know what you are doing.

type InitUpgradeableContract perm = UpgradeableContract (EmptyContractVersion perm)

type OneShotUpgradeParameters ver = ("currentVersion" :! Version, "newVersion" :! Version, "migrationScript" :! MigrationScriptFrom (VerUStoreTemplate ver), "newCode" :! Maybe SomeUContractRouter, "newPermCode" :! Maybe (SomePermanentImpl (VerPermanent ver)))

Parameters of one-shot upgrade.

Do not construct this value manually, consider using makeOneShotUpgradeParameters.

data DVersion

Specify version if given contract.

Constructors

DVersion Version

upgradeableContract :: forall ver. NiceVersion ver => UpgradeableContract ver

mkEmptyStorage :: (VerInterface ver ~ '[], VerUStoreTemplate ver ~ ()) => Address -> Storage ver

pbsContainedInRun :: ParamBuildingStep

Mentions that parameter should be wrapped into Run entry point.

pbsContainedInRunPerm :: ParamBuildingStep

Mentions that parameter should be wrapped into Run entry point.

Orphan instances

CustomErrorHasDoc "upgContractIsNotPaused"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

CustomErrorHasDoc "upgContractIsPaused"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

CustomErrorHasDoc "upgVersionMismatch"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

Buildable (CustomError "upgContractIsNotPaused")

Instance details

Methods

build :: CustomError "upgContractIsNotPaused" -> Builder

Buildable (CustomError "upgContractIsPaused")

Instance details

Methods

build :: CustomError "upgContractIsPaused" -> Builder

Buildable (CustomError "upgVersionMismatch")

Instance details

Methods

build :: CustomError "upgVersionMismatch" -> Builder