Lorentz.Contracts.Upgradeable.Common.Interface

Description

Type-safe interface for constructing contract upgrades.

Use this module as follows: 1. Construct EpwUpgradeParameters; 2. Use one of the respective functions to convert it to an actual upgrade, one-shot or entrypoint-wise, for tests or production.

Documentation

data EpwUpgradeParameters (t :: Type -> Type) (curVer :: VersionKind) (newVer :: VersionKind)

Type-safe upgrade construction.

Constructors

(Traversable t, KnownContractVersion curVer, KnownContractVersion newVer, RequireSamePermanents (VerPermanent curVer) (VerPermanent newVer), RecognizeUpgPiece (UContractRouterUpdate curVer newVer) code, RecognizeUpgPiece (PermanentImplUpdate curVer newVer) codePerm) => EpwUpgradeParameters

Fields

upMigrationScripts :: t (MigrationScript (VerUStoreTemplate curVer) (VerUStoreTemplate newVer))

Storage migration script. Supply this field with result of migrationToScriptI or migrationToScripts call.

upNewCode :: code

Updated parameter dispatching logic. Pass UContractRouter or Nothing.

upNewPermCode :: codePerm

Updates implementation of permanent part. Pass PermanentImpl or Nothing.

fvUpgrade :: forall ver t. (KnownContractVersion ver, Traversable t) => t (MigrationScript (VerUStoreTemplate ver) (VerUStoreTemplate ver)) -> EpwUpgradeParameters t ver ver

Make up a "fixed version" upgrade. As argument you supply result of migrationToScriptI or migrationToScripts and entrypoint-wise migration will be used inside.

Use this method in case you need to authoritatively perform arbitrary modifications of contract storage.

makeOneShotUpgradeParameters :: forall curVer newVer. EpwUpgradeParameters Identity curVer newVer -> OneShotUpgradeParameters curVer

Construct OneShotUpgradeParameters.

Naturally, you can construct this kind of upgrade only if your migration has exactly one stage; for batched migrations use makeEpwUpgrade.

makeOneShotUpgrade :: forall oldVer newVer. EpwUpgradeParameters Identity oldVer newVer -> Parameter oldVer

Construct a call which should be performed in order to perform migration.

makeEpwUpgrade :: forall curVer newVer t. EpwUpgradeParameters t curVer newVer -> [Parameter curVer]

Construct calls which should be performed in order to perform full entrypoint-wise migration.

data UpgradeWay (t :: Type -> Type) where

Way of performing an upgrade.

Constructors

UpgOneShot :: UpgradeWay Identity
Perform upgrade in a single transaction. This, naturally, cannot be used with batched migrations.

UpgEntryPointWise :: UpgradeWay t
Perform upgrade calling one entrypoint per transaction.

type SimpleUpgradeWay = UpgradeWay Identity

UpgradeWay which can be used with simple (non-batched) migrations.

integrationalTestUpgrade :: PermConstraint curVer => EpwUpgradeParameters t curVer newVer -> UpgradeWay t -> UTAddress curVer -> IntegrationalScenarioM (UTAddress newVer)

Perform a contract upgrade in an integrational test scenario.