Lorentz.Contracts.Upgradeable.EntryPointWise

Documentation

type EntryPointImpl store = Lambda (ByteString, UStore store) ([Operation], UStore store)

A helper type that defines an entrypoint that receives a packed argument, i.e. it's basically an unpack instruction followed by a TypedEntryPoint code

type EpwFallback store = Lambda ((MText, ByteString), UStore store) ([Operation], UStore store)

A helper type that defines a function being called in case no implementation matches the requested entrypoint

data EpwContract ver

This data type represents the new contract code and migrations necessary to upgrade the contract endpoints to the new version.

Constructors

EpwContract

Fields

epwServe :: UContractRouter ver

epwServe does the dispatching logic and is assumed to be used for the code lambda of the upgradeable contract.

epwCodeMigrations :: forall oldStore. [MigrationScript oldStore (VerUStoreTemplate ver)]

epwCodeMigrations is a list of packed migrations the client ought to pass to the EpwUpgrade method in order to upgrade the implementation.

data EpwCaseClause store (entry :: EntryPointKind) where

A data type representing a full case clause with the name and implementation of an entrypoint.

Constructors

EpwCaseClause :: TypedEntryPointImpl arg store -> EpwCaseClause store '(name, arg)

mkEpwContract :: forall (ver :: VersionKind) (interface :: [EntryPointKind]) store. (interface ~ VerInterface ver, store ~ VerUStoreTemplate ver, CodeMigrations interface, HasUStore "code" MText (EntryPointImpl store) store, HasUField "fallback" (EpwFallback store) store) => Rec (EpwCaseClause store) interface -> EpwFallback store -> EpwContract ver

Creates the EpwContract data structure from a Rec of case clauses

mkEpwContractT :: forall clauses ver (interface :: [EntryPointKind]) store. (interface ~ VerInterface ver, store ~ VerUStoreTemplate ver, clauses ~ Rec (EpwCaseClause store) interface, RecFromTuple clauses, CodeMigrations interface, HasUStore "code" MText (EntryPointImpl store) store, HasUField "fallback" (EpwFallback store) store) => IsoRecTuple clauses -> EpwFallback store -> EpwContract ver

Like mkEpwContract, but accepts a tuple of clauses, not a Rec.

epwFallbackFail :: EpwFallback store

Default implementation for EpwFallback reports an error just like its UParam counterpart

(/==>) :: Label name -> Lambda (arg, UStore store) ([Operation], UStore store) -> EpwCaseClause store '(name, arg) infixr 0

removeEndpoint :: forall store name s. GetUStoreKey store "code" ~ MText => Label name -> (UStore store ': s) :-> (UStore store ': s)

Removes an endpoint from the #code submap

class EpwDocumented (entries :: [EntryPointKind]) where

Helper for documenting entrypoints with EPW interface.

Methods

epwDocument :: Rec (EpwCaseClause store) entries -> Lambda () ()

Make up documentation for given entry points.

As result you get a fake contract from which you can later build desired documentation. Although, you may want to add contract name and description first.

epwContractDoc :: forall ver. (KnownContractVersion ver, EpwDocumented (VerInterface ver), TypeHasDoc (VerPermanent ver), HasTypeAnn (VerPermanent ver), KnownValue (VerPermanent ver), PermConstraint ver, Typeable ver) => Rec (EpwCaseClause (VerUStoreTemplate ver)) (VerInterface ver) -> PermanentImpl ver -> Lambda () ()

By given list of entrypoints make up a fake contract which contains documentation for the body of given upgradeable contract.