Lorentz.Contracts.Upgradeable.StorageDriven

Description

Contracts based on storage-driven upgrages.

Here entrypoints are declared as part of UStore template, then we automatically derive public API of the contract from it.

Migration mechanism for storage (see Migration) also applies to these entrypoints.

This approach differs from one in Lorentz.Contracts.Upgradeable.EntryPointWise in the following points: 1. Storage migrations are not handled here, only UContractRouter creation. The former is comprehensively handled by Migration. 2. Contract interface is declared via storage - UStoreEntryPoint entries in storage define which public entrypoints (those which are callable via passing UParam) the contract will have. 3. Parameter dispatch fallback is made part of UContractRouter, not storing it in storage here for simplicity. The user can still decide to keep fallback implementation in storage if it is big and then refer to it in SduFallback.

Documentation

type UStoreEntryPoint arg store = UStoreFieldExt UMarkerEntryPoint (SduEntryPoint arg store)

Public entrypoint of a contract kept in UStore.

These are mere UStore fields but treated specially by mkSduContract function which produces UContractRouter capable of calling these entrypoints.

This type is not intended for keeping internal code, in such case consider using UStoreField instead.

data UMarkerEntryPoint :: UStoreMarkerType

data SduEntryPoint (store :: Type) (arg :: Type)

An entrypoint which is assumed to be kept in UStore. It accepts a packed argument.

mkSduEntryPoint :: NiceUnpackedValue arg => Entrypoint arg (UStore store) -> SduEntryPoint store arg

Construct public entrypoint.

mkUStoreEntryPoint :: NiceUnpackedValue arg => Entrypoint arg (UStore store) -> UStoreEntryPoint store arg

Construct public entrypoint for UStore.

type SduFallback store = Lambda ((MText, ByteString), UStore store) ([Operation], UStore store)

A helper type that defines a function being called in case no implementation matches the requested entrypoint

type UStoreEpInterface utemplate = ExtractInterface utemplate (PickMarkedFields UMarkerEntryPoint utemplate)

Get the set of entrypoints (i.e. UStoreEntryPoint entries) stored in UStore with given template.

mkSduContract :: SduFallback (VerUStoreTemplate ver) -> UContractRouter ver

Construct UContractRouter which allows calling all entrypoints stored as UStoreEntryPoint entries of UStore.

callUStoreEntryPoint :: (NicePackedValue arg, HasUField field (SduEntryPoint store arg) store) => Label field -> (arg ': (UStore store ': s)) :-> (([Operation], UStore store) ': s)

Call an entrypoint stored under the given field.

sduFallbackFail :: SduFallback store

Default implementation for SduFallback reports an error just like its UParam counterpart.

Documentation

sduDocument :: UStoreEpTraversal '[Typeable] template => template -> Lambda () ()

Gather documentation of entrypoints kept in given storage. Unfortunatelly, this seems to be the only place where we can pick the code for documenting it.

Note: in most cases you want to use this function is couple with sduAddEntrypointDoc.

sduAddEntrypointDoc :: (UStoreEpTraversal [KnownValue, TypeHasDoc, HasTypeAnn] template, DocItem (DEntryPoint epKind)) => Proxy epKind -> template -> template

Mark all public code kept in given storage as atomic entrypoints.

Sometimes you want your SduEntryPoints to contain multiple sub-entrypoints inside, in this case using entryCase function you get documentation for each of sub-entrypoints automatically and calling this function is not necessary. In case when this does not hold and SduEntryPoint keeps exactly one entrypoint, you still need to mark it as such in order for sduDocument to handle it properly. This function does exactly that - it finds all UStore entrypoints and marks them for documentation.

sduContractDoc :: forall utemplate ver. (KnownContractVersion ver, UStoreEpTraversal '[Typeable] utemplate, TypeHasDoc (VerPermanent ver), KnownValue (VerPermanent ver), HasTypeAnn (VerPermanent ver), PermConstraint ver, Typeable ver) => utemplate -> PermanentImpl ver -> Lambda () ()

By given storage make up a fake contract which contains documentation of all entrypoints declared by this storage.

Note: in most cases you want to use this function in couple with sduAddEntrypointDoc.

Note: we intentionally allow accepted UStore template not to correspond to the contract version storage, this is useful when one does not want to provide the full storage (construction of which may require passing some parameters), rather only part of storage with entrypoints.

Internals

type UStoreEpTraversal c a = (Generic a, GUStoreEpTraversal c (Rep a))