Lorentz.Contracts.UpgradeableCounter.V1

Documentation

data CounterV1 :: VersionKind

migrate :: MigrationScript () UStoreTemplateV1

migrations :: [MigrationScript () UStoreTemplateV1]

Migrations represent entrypoint-wise upgrades. Each migration puts an implementation of a method to UStore. The contract code itself (epwServe) does not do anything special except for taking these lambdas out of the big map.

counterContract :: UContractRouter CounterV1

counterUpgradeParameters :: EpwUpgradeParameters [] CounterV0 CounterV1

type UStoreV1 = UStore UStoreTemplateV1

data UStoreTemplateV1