Lorentz.Contracts.UpgradeableCounterSdu.V1

Documentation

counterContract :: UContractRouter (CounterSduV 1)

migration :: UStoreMigration () UStoreTemplate

This function migrates the storage from an empty one to UStorage, i.e. it populates the empty BigMap with initial values for each field and entrypoints. The result is expected to adhere to V1.UStoreTemplate.

counterUpgradeParameters :: EpwUpgradeParameters Identity (CounterSduV 0) (CounterSduV 1)

counterDoc :: ContractDoc

counterDocRendered :: LText

Internals

runGetCounterValue :: Entrypoint (Void_ () Natural) UStorage

runAdd :: Entrypoint Natural UStorage

permImpl :: PermanentImpl (CounterSduV 1)

Orphan instances

KnownContractVersion (CounterSduV 1)

Instance details

Associated Types

type VerInterface (CounterSduV 1) :: [EntryPointKind]

type VerUStoreTemplate (CounterSduV 1) :: Type

type VerPermanent (CounterSduV 1) :: Type

Methods

contractVersion :: Proxy (CounterSduV 1) -> Version