Lorentz.Contracts.UpgradeableCounterSdu.V2

Documentation

data CounterSduV (v :: Nat) :: VersionKind

Version identifier for this contract.

It a bit differs from how we do in other contracts - this type is supposed to be used in all versions of the contract, but it has type parameter which is supposed to designate contract version.

counterContract :: UContractRouter (CounterSduV 2)

migration :: UStoreMigration (VerUStoreTemplate (CounterSduV 1)) UStoreTemplate

counterUpgradeParameters :: EpwUpgradeParameters Identity (CounterSduV 1) (CounterSduV 2)

counterUpgradeParametersFromV0 :: EpwUpgradeParameters Identity (CounterSduV 0) (CounterSduV 2)

counterRollbackParameters :: EpwUpgradeParameters Identity (CounterSduV 2) (CounterSduV 1)

Orphan instances

KnownContractVersion (CounterSduV 2)

Instance details

Associated Types

type VerInterface (CounterSduV 2) :: [EntryPointKind]

type VerUStoreTemplate (CounterSduV 2) :: Type

type VerPermanent (CounterSduV 2) :: Type

Methods

contractVersion :: Proxy (CounterSduV 2) -> Version