Lorentz.Contracts.UpgradeableUnsafeLedger

Description

UpgradeableUnsafeLedger demonstrates a real-world use case of an upgradeable contract. Its V1 contains a bug that makes it return an incorrect totalSupply. Fortunately, we can upgrade the contract and fix the bug. This contract does not use entrypoint-wise upgrades. Instead, it upgrades atomically in one transaction. For non-atomic upgrades please see the UpgradeableCounter example.

Documentation

data Parameter (ver :: VersionKind)

Parameter of upgradeable contract. It contains, among others:

Entrypoint for running one of upgradeable entrypoints.

Entrypoint for running one of permanent entrypoints, suitable e.g. for implementing interfaces. 3a. Entrypoint for upgrade in a single call. 3b. Entrypoints for entrypoint-wise upgrade.

Constructors

Run (VerParam ver)

RunPerm (VerPermanent ver)

Upgrade (OneShotUpgradeParameters ver)

GetVersion (View () Version)

SetAdministrator Address

EpwBeginUpgrade ("current" :! Version, "new" :! Version)

EpwApplyMigration (MigrationScriptFrom (VerUStoreTemplate ver))

EpwSetCode SomeUContractRouter

EpwSetPermCode (SomePermanentImpl (VerPermanent ver))

EpwFinishUpgrade

data Storage (ver :: VersionKind)

upgradeableUnsafeLedgerContract :: InitUpgradeableContract Permanent

mkEmptyStorage :: (VerInterface ver ~ '[], VerUStoreTemplate ver ~ ()) => Address -> Storage ver