Lorentz.Contracts.UserUpgradeable.V2

Description

The second version of a minimal user-upgradeable ledger. This version is not designed to be upgraded further — it lacks InitiateMigration and MigrateMyTokens entrypoints. However, it has MigrateFrom (callable from V1), and mints new tokens when a user calls V1.MigrateMyTokens. Other functions (either upgradeability-related or standard Transfer/GetTotalSupply may be added if deemed desirable).

Note that the naming in this module is different from Lorentz.Contracts.Upgradeable: by "migration" here we mean the process of transferring the value from an old contract to the new one rather than applying a transformation to storage. Thus, MigrationScript here is a lambda that forges an operation to migrate user's funds rather than a function that upgrades storage in-place.

Documentation

data Parameter

Constructors

MigrateFrom (Address, Natural)
When called by V1, mints new tokens to Address

GetBalance (View Address Natural)
Returns the balance of a holder.

data Storage

Constructors

Storage

Fields

ledger :: Map Address Natural

We use a Map instead of a BigMap to simplify the implementation a bit.

previousVersion :: Maybe (TAddress Parameter)

migrationTarget :: Maybe MigrationTarget

mkStorage :: TAddress Parameter -> Storage

userUpgradeableContract :: '[(Parameter, Storage)] :-> '[([Operation], Storage)]

Orphan instances

CustomErrorHasDoc "userUpgradable'unauthorizedMigrateFrom"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown